Replaced obfuscation with encryption
The biggest difference is that the salt/iv/key needs to live
in environment variables.  But at least the salt is no longer hard-coded.
This should enable us to run securely in Heroku using their
enviornment variable setting features.

The obfuscate_pwd.rb script was removed and replaced with
an option in birstcl.

Reading the config file now passes through ERB before being parsed with JSON